requirements.txt
tensorflow
keras
scikit-learn
pandas
numpy
requests
dvc[gdrive]
streamlit
plotly
pyyaml
